+++ src/assinafy/resources/__init__.py
@@ -4,6 +4,7 @@ from .documents import DocumentResource
 from .fields import FieldResource
 from .signer_documents import SignerDocumentResource
 from .signers import SignerResource
+from .tags import TagResource
 from .templates import TemplateResource
 from .webhooks import WebhookResource
 
@@ -14,6 +15,7 @@ __all__ = [
     "FieldResource",
     "SignerDocumentResource",
     "SignerResource",
+    "TagResource",
     "TemplateResource",
     "WebhookResource",
 ]